Disable doctest for PlotTools because of exotic error message
On some systems, a "NameError: name 'numpy' is not defined" is raised when running tests using code from PlotTools docstrings.
These docstring tests are duplicate of the test suites in testPlotTools.py, so this removal does not affect test coverage of real code.
It affects test coverage of docstrings, so this issue should still be investigated.