Use 0-based index everywhere for consistency
For the general case when editing arrays, we want to access 0-based indices.
The table headers are already 0-based, so the other n-2 dimensions should be as well.

Make internal value in FrameBrowser equal to what is displayed, for simplicity.

The label still provides 1-based numbers when setting limits with setNFrames
("1 of 10")