[statsWidget] add condition before error log
Make sure the _logger.error("Removing item that is not in table: %s", str(item)) make sense.
We can receive some '_removeItem' signals during zoom 'mouse activation' for example. Because link with the 'sigItemAboutToBeRemoved' signal.
There is no reason to receive errors message in this case for example. So we now check the kinf of item before logging it as an error.